V2.2.8

+++ app/Http/Controllers/API/Owners/UpdatePwdController.php
@@ -45,15 +45,24 @@ class UpdatePwdController extends Controller
         }
     }
 
-
     private function changePassword($request) {
-        $email = DB::table('password_resets')->select('email')->where('token', $request->pinCode)->pluck('email');
-        $owner = Owner::whereEmail($email)->first();
-        $owner->update([
-            'password'=>bcrypt($request->password)
-        ]);
-        $this->getPinCode($request)->delete();
-        return $this->returnSuccessMessage('Owner Password Changed Successfully');
+        try {
+            $email = DB::table('password_resets')->select('email')->where('token', $request->pinCode)->pluck('email');
+            if (count($email) > 0 )
+            {
+                $owner = Owner::whereEmail($email)->first();
+                $owner->update([
+                    'password'=>bcrypt($request->password)
+                ]);
+                $this->getPinCode($request)->delete();
+                return $this->returnSuccessMessage('Owner Password Changed Successfully');
+            }else
+            {
+                return $this->returnError('E001', 'PIN Code Not Attached to any email ! check PIN Code');
+            }
+        }catch (\Exception $ex) {
+            return $this->returnError($ex->getCode(), $ex->getMessage());
+        }
     }
 
     private function getPinCode($request){
